+++ internal/features/request-teknisi/entity/entityCore.go
@@ -5,12 +5,12 @@ import (
 )
 
 type Core struct {
-	Id                  int
+	Id                  uint
 	UsersID             uint
 	TeknisiID           uint
 	Foto                []FotoCore
 	Deskripsi           string
-	Jarak               int
+	Jarak              float64
 	Biaya               float64
 	Diproses            bool
 	Konfirmasi_biaya    bool
